// src/timesheet-web/src/app/api/models/timesheet-monthly-report-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MonthlyBacklogTaskTotal } from '../models/monthly-backlog-task-total';
import { TeamNode } from '../models/team-node';
export interface TimesheetMonthlyReportResponse {
  monthlyTotals?: Array<MonthlyBacklogTaskTotal> | null;
  projectTree?: Array<TeamNode> | null;
}