// src/timesheet-web/src/app/api/models/timesheet-weekly-report-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DaysLoggedStat } from '../models/days-logged-stat';
import { WeeklyDayTotal } from '../models/weekly-day-total';
import { WeeklyDetailRow } from '../models/weekly-detail-row';
export interface TimesheetWeeklyReportResponse {
  dayTotals?: Array<WeeklyDayTotal> | null;
  daysLogged?: DaysLoggedStat;
  detailRows?: Array<WeeklyDetailRow> | null;
}